# maas-service/docs/swagger.yaml
definitions:
...
  model.Classifier:
    properties:
      name:
        type: string
      namespace:
        type: string
      tenantId:
        type: string
    required:
    - name
    - namespace
    type: object
...
  model.KafkaCredentials:
    properties:
      clientCert:
        type: string
      clientKey:
        type: string
      password:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
...
  model.KafkaTopicSyncReport:
    properties:
      classifier:
        $ref: '#/definitions/model.Classifier'
      errMsg:
        example: can not get topic info
        type: string
      name:
        example: maas.local.test-topic-0
        type: string
      status:
        example: added
        type: string
    type: object
...
  model.TopicDeletionError:
    properties:
      message:
        example: <error message>
        type: string
      topic:
        $ref: '#/definitions/model.TopicRegistrationRespDto'
    type: object
  model.TopicDeletionResp:
    properties:
      deletedSuccessfully:
        items:
          $ref: '#/definitions/model.TopicRegistrationRespDto'
        type: array
      failedToDelete:
        items:
          $ref: '#/definitions/model.TopicDeletionError'
        type: array
    type: object
  model.TopicRegistrationRespDto:
    properties:
      actualSettings:
        $ref: '#/definitions/model.TopicSettings'
      addresses:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      caCert:
        example: ' '
        type: string
      classifier:
        $ref: '#/definitions/model.Classifier'
      credential:
        additionalProperties:
          items:
            $ref: '#/definitions/model.KafkaCredentials'
          type: array
        type: object
      externallyManaged:
        type: boolean
      instance:
        example: localkafka
        type: string
      name:
        example: local.sampleTopic2.430d243ba9c74f25bf05f492a52804fd
        type: string
      namespace:
        example: local
        type: string
      requestedSettings:
        $ref: '#/definitions/model.TopicSettings'
      template:
        example: "2"
        type: string
      versioned:
        example: false
        type: boolean
    type: object
  model.TopicSearchRequest:
    properties:
      classifier:
        $ref: '#/definitions/model.Classifier'
      instance:
        example: <id of kafka instance where topics are created>
        type: string
      leaveRealTopicIntact:
        type: boolean
      namespace:
        example: <namespace to which topics belong>
        type: string
      template:
        type: integer
      topic:
        example: <exact topic name in kafka>
        type: string
      versioned:
        type: boolean
    type: object
  model.TopicSettings:
    properties:
      configs:
        additionalProperties:
          type: string
        type: object
      minNumPartitions:
        example: 1
        type: integer
      numPartitions:
        example: 1
        type: integer
      replicaAssignment:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
      replicationFactor:
        example: 1
        type: integer
      versioned:
        example: true
        type: boolean
    type: object
  model.TopicTemplate:
    properties:
      configs:
        additionalProperties:
          type: string
        type: object
      minNumPartitions:
        example: 1
        type: integer
      name:
        example: topic Name
        type: string
      namespace:
        example: local
        type: string
      numPartitions:
        example: 1
        type: integer
      replicaAssignment:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
      replicationFactor:
        example: 1
        type: integer
      versioned:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: API for Maas Service.
  title: Maas Service API
paths:
...
  /api/v1/kafka/topic:
    delete:
      description: Delete Topic
      operationId: DeleteTopic
      parameters:
      - description: TopicSearchRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TopicSearchRequest'
      - description: X-Origin-Namespace
        in: header
        name: X-Origin-Namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TopicDeletionResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Delete Topic using Agent Role
      tags:
      - V1
    post:
      description: Get Or Create Topic
      operationId: GetOrCreateTopicV1
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TopicRegistrationRespDto'
      - description: X-Origin-Namespace
        in: header
        name: X-Origin-Namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TopicRegistrationRespDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "410":
          description: Gone
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Get Or Create Topic using Agent Role
      tags:
      - V1
...
  /api/v2/kafka/recover-topic:
    post:
      description: Sync Topic To Kafka
      operationId: SyncTopicToKafka
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.KafkaTopicSyncReport'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Sync Topic To Kafka using Agent Role
      tags:
      - V2
  /api/v2/kafka/recovery/{namespace}:
    post:
      description: Sync All Topics To Kafka
      operationId: SyncAllTopicsToKafka
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.KafkaTopicSyncReport'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Sync All Topics To Kafka using Agent Role
      tags:
      - V2
  /api/v2/kafka/topic:
    post:
      description: Get Or Create Topic
      operationId: GetOrCreateTopicV2
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TopicTemplate'
      - description: X-Origin-Namespace
        in: header
        name: X-Origin-Namespace
        required: true
        type: string
      - description: onTopicExists
        in: query
        name: onTopicExists
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TopicRegistrationRespDto'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TopicRegistrationRespDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "410":
          description: Gone
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Get Or Create Topic using Agent Role
      tags:
      - V2
...
  /api/v2/kafka/topic/watch-create:
    post:
      description: Watch Topics Create
      operationId: WatchTopicsCreate
      parameters:
      - description: X-Origin-Namespace
        in: header
        name: X-Origin-Namespace
        required: true
        type: string
      - description: timeout
        in: query
        name: timeout
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TopicRegistrationRespDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Watch Topics Create using Agent Role
      tags:
      - V2
...
produces:
- application/json
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"